+++ src/services/Fields.php
@@ -655,31 +655,13 @@ SQL
                 ]);
                 $ownershipData[] = [$blockRevisionId, $revision->id, $block->sortOrder];
 
-                // Get the actual blocks, for block structure creation
-                if ($blockRevisionId === $block->id) {
-                    $jobData[$siteId][] = [
-                        'id' => $block->id,
-                        'lft' => $block->lft,
-                        'rgt' => $block->rgt,
-                        'level' => $block->level,
-                    ];
-                } else {
-                    // Querying the database because `getElementById()` doesn't seem to return anything at this point
-                    $jobData[$siteId][] = (new Query())
-                        ->select([
-                            'id' => 'elements.id',
-                            'lft' => 'structureelements.lft',
-                            'rgt' => 'structureelements.rgt',
-                            'level' => 'structureelements.level',
-                        ])
-                        ->from(['elements' => '{{%elements}}'])
-                        ->innerJoin(
-                            ['structureelements' => '{{%structureelements}}'],
-                            '[[structureelements.elementId]] = [[elements.canonicalId]]',
-                        )
-                        ->where(['elements.id' => $blockRevisionId])
-                        ->one();
-                }
+                // Store the job data for block structure creation
+                $jobData[$siteId][] = [
+                    'id' => $blockRevisionId,
+                    'level' => $block->level,
+                    'lft' => $block->lft,
+                    'rgt' => $block->rgt,
+                ];
             }
 
             foreach ($supportedSites ?? [] as $supportedSite) {